step4: 지뢰 찾기(리팩토링) 피드백 반영 - indexSquare 함수를 MineSweeperMap 클래스에서 하도록 변경
